==> 05_RLQ.Rmd <==
---
title: "RLQ"
output: html_document
date: "2024-01-27"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(ade4)
library(tidyverse)
library(vegan)
library(ggplot2)
library(ggrepel)
library(gridExtra)
library(RColorBrewer)
library(distillR)
```

```{r }
load("resources/data/data.Rdata")
```

```{r}
#Aggregate bundle-level GIFTs into the compound level
GIFTs_elements <- to.elements(genome_gifts,GIFT_db)
GIFTs_elements_filtered <- GIFTs_elements[rownames(GIFTs_elements) %in% genome_counts$genome,]
GIFTs_elements_filtered <- as.data.frame(GIFTs_elements_filtered) %>% 
  select_if(~ !is.numeric(.) || sum(.) != 0)

#Aggregate element-level GIFTs into the function level
GIFTs_functions <- to.functions(GIFTs_elements_filtered,GIFT_db)
functions <- GIFTs_functions %>%
  as.data.frame()

#Aggregate function-level GIFTs into overall Biosynthesis, Degradation and Structural GIFTs
GIFTs_domains <- to.domains(GIFTs_functions,GIFT_db)
domains <- GIFTs_domains %>%
  as.data.frame()
#Get community-weighed average GIFTs per sample
genome_counts_row <- genome_counts %>% column_to_rownames(., "genome")
GIFTs_elements_community <- to.community(GIFTs_elements_filtered,genome_counts_row,GIFT_db)
GIFTs_functions_community <- to.community(GIFTs_functions,genome_counts_row,GIFT_db)
GIFTs_domains_community <- to.community(GIFTs_domains,genome_counts_row,GIFT_db)

```

```{r elements, comment="", echo=FALSE, message=FALSE, warning=FALSE}
mag_weighted <- round(
  sweep(genome_counts_row, MARGIN = 1, 1000, `*`), 0
)

mag_weighted_t <-subset(mag_weighted, rownames(mag_weighted) %in% rownames(functions))%>%
  t() %>%
#  row_to_names(row_number = 1) %>%
  as.data.frame()  %>%
  rownames_to_column(., "samples")  %>%
  dplyr::arrange(samples) %>%
  column_to_rownames(., "samples")

```


```{r}
# metadata binary
design <- sample_metadata[, c(
  "sample", "region", "sex"
)]
design$Region<-0
design$Region[design$region=="Daneborg"]<-1
for(i in 1:ncol(design)){
  if(is.character(design[,i])){
    design[,i]<-factor(design[,i])
  }
}
design$Sex<-0
design$sex[design$sex=="Female"]<-1
for(i in 1:ncol(design)){
  if(is.character(design[,i])){
    design[,i]<-factor(design[,i])
  }
}
```

# RLQ analysis (functions)

The multivariate RLQ analysis relates a species x traits matrix (Q, species are rows and columns are biological traits) to a matrix of 
environmental variables at each site (R, sites are rows and columns are environmental descriptors), using a species x abundance matrix 
(L, sites are rows and columsn are abundances of specific species) as a link.

```{r}
genome_counts_t_rel <- decostand(mag_weighted_t,method = "total")
genome_counts_t_log <- log1p(genome_counts_t_rel)
```


```{r}
four <- fourthcorner(data.frame(design[c(4)]), genome_counts_t_log, functions, modeltype = 6, 
                     nrepet = 999,p.adjust.method.G = "fdr",p.adjust.method.D = "fdr")
summary(four)

four_domains <- fourthcorner(data.frame(design[c(4)]), genome_counts_t_log, domains, modeltype = 6, 
                     nrepet = 999,p.adjust.method.G = "fdr",p.adjust.method.D = "fdr")
summary(four_domains)
```









A preliminary step of RLQ analysis is to perform the separate analyses of each table
```{r}
#Correspondence analysis (dudi.coa) is applied to the species table.
AnalysisL_functions <- dudi.coa(genome_counts_t_log, scannf = FALSE)
#Separate analyses of traits and environmental variables should be weighted by the sites and species weights derived from the previous correspondence analysis.
#Principal Component Analysis (PCA) if all variables are continuous
AnalysisQ_functions <- dudi.pca(functions, row.w = AnalysisL_functions$cw,scannf = FALSE)
#Hill and smith analysis if there is a mix of continuous and categorical variables. Hill and smith analysis is calculated with dudi.hillsmith()
AnalysisR_functions <- dudi.hillsmith(design[c(4,5)], row.w = AnalysisL_functions$lw,scannf = FALSE)#function that allows to consider mix of different types of variables
```

RLQ is based on co-inertia analysis, which is a unconstrained symmetrical analysis that look for a compromise between the axes of three unconstrained analyses. RLQ combines the three separate analyses and aims at identifying the main relationships between environmental gradients and trait syndromes mediated by species abundances.
```{r}
rlq.functions <- rlq(AnalysisR_functions, AnalysisL_functions, AnalysisQ_functions,scannf = FALSE)
summary(rlq.functions)
```
The projected inertia is the amount of co-inertia explained by the the ordination provided by RLQ analysis assigns scores to species, 
samples, traits, and environmental variables along orthogonal axes and yields graphical summary of the main structures.
The first axis of co-inertia (RLQ1) explain 43% of covariance. Hence, we will only interpret the scores of the first axis (but similar visualization could be carried out for subsequent axis).

#### Combining with fourth corner

```{r}
set.seed(111)
testrlq.functions <- randtest(rlq.functions, modeltype = 6, nrepet = 999)
testrlq.functions
plot(testrlq.functions)
```

***Both models need to be significant to continue***
```{r}


# If the models are significant, continue with the analyses
set.seed(111)
testQaxes.comb.functions <- fourthcorner.rlq(rlq.functions, modeltype = 6,
                                            typetest = "Q.axes", nrepet = 99, p.adjust.method.G = "fdr",
                                            p.adjust.method.D = "fdr")
testRaxes.comb.functions <- fourthcorner.rlq(rlq.functions, modeltype = 6,
                                            typetest = "R.axes", nrepet = 99, p.adjust.method.G = "fdr",
                                            p.adjust.method.D = "fdr")
print(testQaxes.comb.functions, stat = "D")
print(testRaxes.comb.functions, stat = "D")
```

#### Plotting
*** Significant positive relationships are depicted with red and, negative relationships with blue.***
```{r}
par(mfrow = c(1, 2))
plot(testQaxes.comb.functions, alpha = 0.05, type = "table",
     stat = "D2")
plot(testRaxes.comb.functions, alpha = 0.05, type = "table",
     stat = "D2")
```

*** Blue color indicates significant relationship with the first axis. Orange color indicates significant relationship with second axis. Green indicates significant relationship with both axes. Grey indicates no significant relationship.***
```{r}
gg.functionsQ<-ggplot(rlq.functions$c1,aes(y=CS2,x=CS1))+
  xlim(c(-1,1))+ylim(c(-1,1))+
  geom_hline(yintercept=0,linetype="dashed")+geom_vline(xintercept = 0,linetype="dashed")+
  geom_label_repel(aes(y=CS2,x=CS1),show.legend = FALSE,label=rownames(rlq.functions$c1)) +
  theme_bw()
gg.functionsR<-ggplot(rlq.functions$l1,aes(y=RS2,x=RS1))+
  xlim(c(-1,1))+ylim(c(-1,1))+
  geom_hline(yintercept=0,linetype="dashed")+geom_vline(xintercept = 0,linetype="dashed")+
  geom_label_repel(aes(y=RS2,x=RS1),label=rownames(rlq.functions$l1), show.legend = FALSE) +
  theme_bw()
grid.arrange(gg.functionsQ,gg.functionsR,ncol=2,top="Sample type")

```
